Add files via upload
This contains a updated tracking meditation page created with the help of youTube tutorials and bootstrap with a chart to add notes and stats. A new background and a navigation bar have been added as well to jump between pages. Navigation bar contains home, tracking, timer, blog, that will jump between test1-3 and trackMed html pages to show the ability to link pages.