Updated tracking page with add functionality
The add button opens up a prompt that now allows a user to input into three fields and press "Add" to then append a row to the chart. It will close the prompt at the same time and clear the fields for the next time you add a new note.